Extract workload catalog into a standalone ct-workload crate
The carrier-generic op fns + FixtureCarrier move to a no_std ct-workload crate
so the DWT hardware suite (no_std, size-constrained) can call the same ops
without dragging ct-fixtures' 496 extern-C symbols or its panic-handler/ctgrind
std gating into the firmware. ct-fixtures re-exports it as `catalog` and keeps
the emit_wl_* adapters (which generate the fixtures). Adds the hardware-only
ops (ct_checked_shl/pow via a helper trait, nct_div, nct_ilog10). Gate
unchanged (496/0).
